+++ app/models/batch_connect/session.rb
@@ -1,5 +1,3 @@
-require "refinements/to_bool"
-
 module BatchConnect
   class Session
     include ActiveModel::Model
@@ -8,8 +6,6 @@ module BatchConnect
     # This class describes the object that is bound to the ERB template file
     # when it is rendered
     TemplateBinding = Struct.new(:session, :context) do
-      using Refinements::ToBool
-
       # Get the binding for this object
       # @return [Binding] this object's binding
       def get_binding