+++ app/models/batch_connect/session_context.rb
@@ -1,9 +1,5 @@
-require "refinements/to_bool"
-
 module BatchConnect
   class SessionContext
-    using Refinements::ToBool
-
     include Enumerable
 
     include ActiveModel::Model